Store and event listeners
Cleaned up event listeners so they stopped repeating on hot reload.

Added grid to the store so we can access and change it from multiple mods. Still haven't changed modifications from direct to mutation based. This is honestly pretty low priority for this project.

Backspace now deletes notes if there is no value present, and adding the same value as the current value in the cell deletes the value similarly to how other sudoku applications work.

This is (read: should be) the final commit before starting the solver.